move check to napari.__init__
