+++ napari/__main__.py
@@ -568,38 +568,7 @@ def _maybe_rerun_with_macos_fixes():
                 shutil.rmtree(tempdir)
 
 
-def _check_installation_path():  # pragma: no cover
-    """Check for installation path conflicts.
-
-    Check if napari is present in site-packages. If napari is installed in editable mode,
-    notify the user of a the conflict that napari is also in site-packages.
-    """
-    napari_installation_path = Path(__file__).absolute().parent.parent
-    if napari_installation_path.name == 'site-packages':
-        # napari is installed in non-editable mode
-        return
-
-    import numpy as np
-
-    site_packages_path = Path(np.__file__).absolute().parent.parent
-    if site_packages_path.name != 'site-packages':
-        # numpy is not installed in site-packages
-        return
-
-    problematic_napari_path = site_packages_path / 'napari'
-    if problematic_napari_path.exists():
-        print(  # noqa: T201
-            f'Found a napari directory: {problematic_napari_path}, '
-            'but napari is installed in editable mode. '
-            'Please remove napari directory from site-packages.',
-            file=sys.stderr,
-        )
-        raise SystemExit(1)
-
-
 def main():
-    _check_installation_path()
-
     # There a number of macOS issues we can fix with env vars
     # and/or relaunching a subprocess
     _maybe_rerun_with_macos_fixes()
